requirements.txt
python-telegram-bot==20.6
rasa==3.6.8
aiohttp==3.9.5
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.30
